Move duplicated, move more constructor tests
